df/uucore: refactor - move duplicate code to uucore/fsext.rs
