+++ src/uucore/src/lib/lib.rs
@@ -189,6 +189,7 @@ mod tests {
         vec.into_iter().collect_str(handling)
     }
 
+    #[cfg(any(unix, target_os = "redox"))]
     fn test_invalid_utf8_args_lossy(os_str: &OsStr) {
         //assert our string is invalid utf8
         assert!(os_str.to_os_string().into_string().is_err());
@@ -212,6 +213,7 @@ mod tests {
         );
     }
 
+    #[cfg(any(unix, target_os = "redox"))]
     fn test_invalid_utf8_args_ignore(os_str: &OsStr) {
         //assert our string is invalid utf8
         assert!(os_str.to_os_string().into_string().is_err());
@@ -236,7 +238,7 @@ mod tests {
         //create a vector containing only correct encoding
         let test_vec = make_os_vec(&OsString::from("test2"));
         //expect complete conversion without losses, even when lossy conversion is accepted
-        let _ = collect_os_str(test_vec.clone(), InvalidEncodingHandling::ConvertLossy)
+        let _ = collect_os_str(test_vec, InvalidEncodingHandling::ConvertLossy)
             .expect_complete("Lossy conversion not expected in this test");
     }
